requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
Brotli==1.0.9
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
markdownify==0.11.6
multidict==6.0.3
pycares==4.2.2
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.0
python-telegram-bot @ git+https://github.com/python-telegram-bot/python-telegram-bot@f8be97c12c77ca36b7a5a39df8d68e84429c58f3
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyyml==0.0.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tornado==6.1
tzdata==2022.7
tzlocal==4.2
yarl==1.8.2
